Refactor Slotmap + Handles
Handles are now simply uint32_t. Complexity can be readded later.

For now though, means that inner structs of Renderer do not need to be exposed. The engine can then have a GameScene class that has these objects and is able to use them without needing to have vulkan exposed to it :)